backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
httpx==0.27.2
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
pyyaml==6.0.2
